Struct actix_settings::ActixSettings

pub struct ActixSettings {

    pub hosts: Vec<Address>,
    pub mode: Mode,
    pub enable_compression: bool,
    pub enable_log: bool,
    pub num_workers: NumWorkers,
    pub backlog: Backlog,
    pub max_connections: MaxConnections,
    pub max_connection_rate: MaxConnectionRate,
    pub keep_alive: KeepAlive,
    pub client_timeout: Timeout,
    pub client_shutdown: Timeout,
    pub shutdown_timeout: Timeout,
    pub tls: Tls,

}

Settings types for Actix Web.

Fields
hosts: Vec<Address>

List of addresses for the server to bind to.

mode: Mode

Marker of intended deployment environment.

enable_compression: bool

True if the Compress middleware should be enabled.

enable_log: bool

True if the Logger middleware should be enabled.

num_workers: NumWorkers

The number of workers that the server should start.

backlog: Backlog

The maximum number of pending connections.

max_connections: MaxConnections

The per-worker maximum number of concurrent connections.

max_connection_rate: MaxConnectionRate

The per-worker maximum concurrent TLS connection limit.

keep_alive: KeepAlive

Server keep-alive preference.

client_timeout: Timeout

Timeout duration for reading client request header.

client_shutdown: Timeout

Timeout duration for connection shutdown.

shutdown_timeout: Timeout

Timeout duration for graceful worker shutdown.

tls: Tls

TLS (HTTPS) configuration.