Struct actix_settings::BasicSettings

pub struct BasicSettings<A> {
    pub actix: ActixSettings,
    pub application: A,
}

Wrapper for server and application-specific settings.

Fields
actix: ActixSettings

Actix Web server settings.

application: A

Application-specific settings.

Implementations

impl<A> BasicSettings<A>where A: DeserializeOwned,

pub fn parse_toml<P>(filepath: P) -> Result<Self, Error>where P: AsRef<Path>,

Parse an instance of Self from a TOML file located at filepath.

If the file doesn’t exist, it is generated from the default TOML template, after which the newly generated file is read in and parsed.

pub fn from_default_template() -> Result<Self, Error>

Parse an instance of Self straight from the default TOML template.

pub fn from_template(template: &str) -> Result<Self, Error>

Parse an instance of Self straight from the default TOML template.

pub fn write_toml_file<P>(filepath: P) -> Result<(), Error>where P: AsRef<Path>,

Writes the default TOML template to a new file, located at filepath.

Errors

Returns a FileExists error if a file already exists at that location.

pub fn override_field<F, V>(field: &mut F, value: V) -> Result<(), Error>where F: Parse, V: AsRef<str>,

Attempts to parse value and override the referenced field.

Examples

use actix_settings::{Settings, Mode};

let mut settings = Settings::from_default_template()?;
assert_eq!(settings.actix.mode, Mode::Development);

Settings::override_field(&mut settings.actix.mode, "production")?;
assert_eq!(settings.actix.mode, Mode::Production);

pub fn override_field_with_env_var<F, N>( field: &mut F, var_name: N ) -> Result<(), Error>where F: Parse, N: AsRef<str>,

Attempts to read an environment variable, parse it, and override the referenced field.

Examples

use actix_settings::{Settings, Mode};

std::env::set_var("OVERRIDE__MODE", "production");

let mut settings = Settings::from_default_template()?;
assert_eq!(settings.actix.mode, Mode::Development);

Settings::override_field_with_env_var(&mut settings.actix.mode, "OVERRIDE__MODE")?;
assert_eq!(settings.actix.mode, Mode::Production);